create useTransaction.ts + error fixing
